fix(command.ts): align order of variables with error message
I spent a lot of time debugging failed snapshot tests until I found out that the
variables for expected and observed image dimensions are swapped in the error message.
Can be tested by updating the snapshot and then checking which values change in the message.
Fix is to just put them in the correct order.